BASE-4240: report_coursesize: Update "last calculated" timestamp after calculation
Without this, users see the 'last calculated' timestamp in the UI updated
before the calculations actually complete. This becomes a bigger problem when
the calculations themselves die due to an issue somewhere and they don't
actually get performed for a long time - giving users an outdated set of data
they think is current.

Change-Id: I289407d3ca3276403aeaea65e95bae844b0115cc
Reviewed-on: https://gerrit-ui.service.shr-syd.mretech.io/c/development/moodle/+/39144
Reviewed-by: Tim Lock <tim.lock@openlms.net>
Tested-by: Jenkins