BASE-4310: report_coursesize: Optimise category calculation query
Rather than pull in courseids to PHP only to send back to the DB as query
params - its more optimal to let the db manage it entirely.

Especially since, the number of IDs is unbounded - and at a certain scale ends
up exceeding the number of query params we can send to the DB for a single
query.

Change-Id: I957d0db3de1abce7756d38125a82a569fc16ce5d
Reviewed-on: https://gerrit-ui.service.shr-syd.mretech.io/c/development/moodle/+/39386
Tested-by: Jenkins
Reviewed-by: Tim Lock <tim.lock@openlms.net>